+++ ecc/bls12-377/fr/mimc/mimc.go
@@ -21,7 +21,6 @@ import (
 	"hash"
 
 	"github.com/consensys/gnark-crypto/ecc/bls12-377/fr"
-	"golang.org/x/crypto/sha3"
 	"math/big"
 	"sync"
 )
@@ -34,8 +33,7 @@ const (
 
 // Params constants for the mimc hash function
 var (
-	mimcConstants [mimcNbRounds]fr.Element
-	once          sync.Once
+	once sync.Once
 )
 
 // digest represents the partial evaluation of the checksum
@@ -48,10 +46,9 @@ type digest struct {
 
 // GetConstants exposed to be used in gnark
 func GetConstants() []big.Int {
-	once.Do(initConstants) // init constants
 	res := make([]big.Int, mimcNbRounds)
 	for i := 0; i < mimcNbRounds; i++ {
-		mimcConstants[i].BigInt(&res[i])
+		fr.MIMCConstants[i].BigInt(&res[i])
 	}
 	return res
 }
@@ -75,14 +72,6 @@ func (d *digest) Reset() {
 // It does not change the underlying hash state.
 func (d *digest) Sum(b []byte) []byte {
 	buffer := d.checksum()
-	d.data = nil // flush the data already hashed
-	hash := buffer.Bytes()
-	b = append(b, hash[:]...)
-	return b
-}
-
-func (d *digest) Sum2(b []byte) []byte {
-	buffer := d.checksum2()
 	d.data = nil // flush the data already hashed
 	hash := buffer.Bytes()
 	b = append(b, hash[:]...)
@@ -155,45 +144,10 @@ func (d *digest) checksum() fr.Element {
 	return d.h
 }
 
-func (d *digest) checksum2() fr.Element {
-	// Write guarantees len(data) % BlockSize == 0
-
-	// TODO @ThomasPiellard shouldn't Sum() returns an error if there is no data?
-	// TODO: @Tabaie, @Thomas Piellard Now sure what to make of this
-	/*if len(d.data) == 0 {
-		d.data = make([]byte, BlockSize)
-	}*/
-
-	for i := range d.data {
-		r := d.encrypt2(d.data[i])
-		d.h.Add(&r, &d.h).Add(&d.h, &d.data[i])
-	}
-
-	return d.h
-}
-
 // plain execution of a mimc run
 // m: message
 // k: encryption key
 func (d *digest) encrypt(m fr.Element) fr.Element {
-	once.Do(initConstants) // init constants
-
-	var tmp fr.Element
-	for i := 0; i < mimcNbRounds; i++ {
-		// m = (m+k+c)^**17
-		tmp.Add(&m, &d.h).Add(&tmp, &mimcConstants[i])
-		m.Square(&tmp).
-			Square(&m).
-			Square(&m).
-			Square(&m).
-			Mul(&m, &tmp)
-	}
-	m.Add(&m, &d.h)
-	return m
-}
-
-func (d *digest) encrypt2(m fr.Element) fr.Element {
-	once.Do(initConstants) // init constants
 	fr.MIMCEncrypt(&d.h, &m)
 	return m
 }
@@ -207,23 +161,6 @@ func Sum(msg []byte) ([]byte, error) {
 	h := d.checksum()
 	bytes := h.Bytes()
 	return bytes[:], nil
-}
-
-func initConstants() {
-	bseed := ([]byte)(seed)
-
-	hash := sha3.NewLegacyKeccak256()
-	_, _ = hash.Write(bseed)
-	rnd := hash.Sum(nil) // pre hash before use
-	hash.Reset()
-	_, _ = hash.Write(rnd)
-
-	for i := 0; i < mimcNbRounds; i++ {
-		rnd = hash.Sum(nil)
-		mimcConstants[i].SetBytes(rnd)
-		hash.Reset()
-		_, _ = hash.Write(rnd)
-	}
 }
 
 // WriteString writes a string that doesn't necessarily consist of field elements
