+++ ecc/bls12-377/fr/mimc/mimc_test.go
@@ -1,70 +1,9 @@
 package mimc
 
 import (
-	"bytes"
-	"fmt"
 	"github.com/consensys/gnark-crypto/ecc/bls12-377/fr"
-	"github.com/stretchr/testify/assert"
-	mrand "math/rand"
 	"testing"
 )
-
-func TestHashASMMIMC(t *testing.T) {
-	initConstants()
-	fmt.Println("Running test")
-
-	for i := 0; i < 100; i++ {
-		msg := make([]byte, 128)
-		_, err := mrand.Read(msg)
-		if err != nil {
-			panic(err)
-		}
-
-		h1 := NewMiMC()
-		h2 := NewMiMC()
-
-		res1 := h1.Sum(msg)
-
-		res2 := h2.(*digest).Sum2(msg)
-
-		assert.Equal(t, bytes.Compare(res1, res2), 0)
-	}
-
-}
-
-//func BenchmarkMIMCNative(b *testing.B) {
-//	for i := 0; i < b.N; i++ {
-//		b.StopTimer()
-//		msg := make([]byte, 128)
-//		_, err := mrand.Read(msg)
-//		if err != nil {
-//			panic(err)
-//		}
-//
-//		h := NewMiMC()
-//
-//		b.StartTimer()
-//		_ = h.Sum(msg)
-//		_ = h.Sum(nil)
-//	}
-//}
-//
-//func BenchmarkMIMCAsm(b *testing.B) {
-//	for i := 0; i < b.N; i++ {
-//		b.StopTimer()
-//		msg := make([]byte, 128)
-//		_, err := mrand.Read(msg)
-//		if err != nil {
-//			panic(err)
-//		}
-//
-//		h := NewMiMC()
-//
-//		b.StartTimer()
-//		_ = h.(*digest).Sum2(msg)
-//		_ = h.(*digest).Sum2(nil)
-//	}
-//}
 
 func BenchmarkEncryptASM(b *testing.B) {
 	m, _ := new(fr.Element).SetRandom()
@@ -83,7 +22,6 @@ func BenchmarkEncryptASM(b *testing.B) {
 }
 
 func BenchmarkEncryptSemiASM(b *testing.B) {
-	once.Do(initConstants) // init constants
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
@@ -96,18 +34,15 @@ func BenchmarkEncryptSemiASM(b *testing.B) {
 }
 
 func MIMCEncryptSemi(h, m *fr.Element) {
-	once.Do(initConstants) // init constants
 
 	for i := 0; i < mimcNbRounds; i++ {
-		fr.MIMCStep(h, m, &mimcConstants[i])
+		fr.MIMCStep(h, m, &fr.MIMCConstants[i])
 	}
 
 	m.Add(m, h)
 }
 
 func BenchmarkEncryptNative(b *testing.B) {
-	once.Do(initConstants) // init constants
-
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
